Find missing unix account configs in both files
Previously the test compared the two lists of users to check simply if they're
the same. I've modified the test to compare A to B and B to A in order to find
out which list a user is missing from, if it is missing. It also checks that
the keys are the same.

This should help speed up diagnosing a problem in this config for future devs.